pyproject.toml
[project]
name = "jobs-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.35.0",
    "bs4>=0.0.2",
    "fastapi>=0.122.0",
    "httpx>=0.27.0",  # Required for FastAPI TestClient
    "langchain>=1.0.7",
    "mangum>=0.18.0",
    "openai>=2.8.1",
    "pydantic>=2.12.4",
    "pytest>=9.0.1",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.38.0",
]
